Separate components from <ExpandedStoryDescription />
