+++ app/assets/javascripts/components/story/ExpandedStory/ExpandedStoryDescription/ExpandedStoryDescriptionContent.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Markdown from '../../../Markdown';
+import PropTypes from 'prop-types';
+
+const ExpandedStoryDescriptionContent = ({ description }) =>
+  <div className='markdown-wrapper'>
+    <Markdown source={description} />
+  </div>
+
+ExpandedStoryDescriptionContent.defaultProps = {
+  description: ''
+};
+
+ExpandedStoryDescriptionContent.propTypes = {
+  description: PropTypes.string.isRequired
+};
+
+export default ExpandedStoryDescriptionContent;
