+++ app/assets/javascripts/components/story/ExpandedStory/ExpandedStoryDescription/ExpandedStoryDescriptionEditButton.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const ExpandedStoryDescriptionEditButton = () =>
+  <button className='edit-description-button'>
+    { I18n.t('edit') }
+  </button>
+
+export default ExpandedStoryDescriptionEditButton;
